requirements.txt
google-cloud-bigquery
PyYAML==6.0
